zsh: Add support for xterm-24bit
This solves an issue with emacs in tty not loading the theme properly
when started as a daemon from systemd. For some reason, running it as
daemon directly does not reproduce the issue.
Setting TERM to xterm-24bit solves the issue for systemd, but there's
likely another way since truecolor support is supposed to be built-in
emacs 27.1+...